+++ src/Symfony/Component/ErrorHandler/Tests/Fixtures/RootNamespace.php
@@ -1,0 +1,7 @@
+<?php
+
+use Symfony\Component\ErrorHandler\Tests\Fixtures\DeprecatedClass;
+
+class RootNamespace extends DeprecatedClass
+{
+}